remove hive-connector dep on table-planner internal class
