+++ flink-table/flink-table-planner/src/main/scala/org/apache/flink/table/planner/delegation/PlannerBase.scala
@@ -110,7 +110,7 @@ abstract class PlannerBase(
 
   @VisibleForTesting
   private[flink] val plannerContext: PlannerContext =
-    new PlannerContext(
+    new PlannerContextImpl(
       !isStreamingMode,
       tableConfig,
       moduleManager,
